Experimental split-screen support (#14)
Page component refactor and experimental split-screen support
