+++ src/static/ui.css
@@ -57,7 +57,7 @@ body {
     flex: 1;
     flex-direction: column;
     padding: 0 0 10px 0; }
-  #chat-root .chat-room-sidebar {
+  #chat-root .chat-room-details {
     flex-basis: 240px;
     padding: 0 0 0 20px; }
   #chat-root .chat-room-content-footer {
