+++ src/static/ui.scss
@@ -78,7 +78,7 @@ body {
     padding: 0 0 10px 0;
   }
 
-  .chat-room-sidebar {
+  .chat-room-details {
     flex-basis: 240px;
     padding: 0 0 0 20px;
   }
